fixed classifier type and consolidated unit tests
